pyproject.toml
[project]
name = "pathfinder-discord-bot"
version = "0.1.0"
description = "Pathfinder 2E Discord bot with dice rolling, LLM Q&A, and campaign lore tracking"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "discord.py>=2.3.0",
    "anthropic>=0.39.0",         # Claude API
    "openai>=1.54.0",            # OpenAI embeddings
    "motor>=3.6.0",              # Async MongoDB driver
    "pymongo>=4.10.0",           # Sync MongoDB (for migrations/scripts)
    "pymupdf>=1.24.0",           # PDF parsing
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
    "python-dotenv>=1.0.0",
    "aiohttp>=3.10.0",
    "tenacity>=9.0.0",           # Retry logic for API calls
    "numpy>=2.0.0",              # For vector operations
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.7.0",               # Linting + formatting
    "mypy>=1.13.0",              # Type checking
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
# This tells Hatch to include your main Python package folder in the build.
packages = ["pathfinder_discord_bot"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]  # Line length handled by formatter

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]